Calculate robustness

Uses a maximum flow / minimum cut algorithm (imported from igraph) to calculate the maximum flow / minimum cut from each of the factors for which `found_from` is true to each of the factors for which `found_to` is true.

calculate_robustness() is used by pipe_trace_paths().

pipe_calculate_robustness(graf, field = NULL)

Arguments

graf
A tidymap. To use this function, the factors table of this map must include two logical variables called `found_from` and `found_to`. If the links table contains an integer column `n`, these values are treated as the capacity of the links, otherwise the capacity for each link is taken as 1. If there are multiple links between any ordered pair of nodes, the links are combined using bundle_links()

field
An optional field by which to split the robustness calculation.

Value

A tidymap with an additional attribute `flow`, a tibble (dataframe) in which the columns are each of the factors for which `found_from` is true (if there is more than one such column, and additional "All sources" column is prepended); and in which the rows are each of the factors for which `found_to` is true (if there is more than one such row, and additional "All targets" column is prepended. The (integer) values in the table represent the maximum flow / minimum cut aka Robustness score from the corresponding source factor (column) to the corresponding target factor (row). The scores for the "All sources" column are calculated by constructing an additional factor as an ultimate source which is connected to the other sources by links of infinite capacity, and likewise for the "All targets" row. If `field` is not NULL, robustness is calculated several times, once for each value of that field. In this case, what is returned is a dataframe which summarises the set of dataframes so that the value of each cell in the returned dataframe is the number of these values for which the robustness value in the corresponden is not zero.

The table is sorted in descending order of the first column.

Examples

cashTransferMap %>% pipe_trace_paths(from="Cash",to="Increa",length=4) %>% pipe_merge_statements %>% pipe_calculate_robustness(field="#SourceID") %>% attr("flow")

#> Joining, by = "label"

#> Joining, by = "label"

#> Warning: The `x` argument of `as_tibble.matrix()` must have unique column names if `.name_repair` is omitted as of tibble 2.0.0.
#> Using compatibility `.name_repair`.

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> Joining, by = "label"

#> $`mjf-7`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            13                 9                 4                0
#>  2 (BF) Incr~             1                 1                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             7                 4                 4                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             5                 5                 1                0
#>  7 (HN) Incr~             1                 1                 1                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             1                 1                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             1                 1                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mkf-8`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            11                11                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             5                 5                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             3                 3                 0                0
#>  7 (HN) Incr~             5                 5                 0                0
#>  8 (BF) Incr~             1                 1                 0                0
#>  9 (IEA) Inc~             1                 1                 0                0
#> 10 (IEA) Inc~             3                 3                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             1                 1                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-9`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             7                 7                 0                0
#>  2 (BF) Incr~             1                 1                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             2                 2                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             5                 5                 0                0
#>  7 (HN) Incr~             3                 3                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-10`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             3                 3                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             3                 3                 0                0
#>  7 (HN) Incr~             1                 1                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             1                 1                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mkf-11`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             6                 6                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             5                 5                 0                0
#>  7 (HN) Incr~             3                 3                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             0                 0                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-12`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             6                 6                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             6                 6                 0                0
#>  7 (HN) Incr~             2                 2                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             1                 1                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-1`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             8                 8                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             3                 3                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             4                 4                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mkf-2`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             8                 8                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             6                 6                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mkf-3`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             9                 7                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             7                 7                 0                0
#>  7 (HN) Incr~             3                 3                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             3                 3                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             2                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-4`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             9                 7                 2                0
#>  2 (BF) Incr~             2                 2                 2                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             9                 7                 2                0
#>  7 (HN) Incr~             5                 5                 2                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             0                 0                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-5`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             7                 7                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             7                 7                 0                0
#>  7 (HN) Incr~             6                 6                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             3                 3                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mkf-6`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            17                10                 7                0
#>  2 (BF) Incr~             2                 2                 2                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 1                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~            16                 9                 7                0
#>  7 (HN) Incr~             7                 7                 7                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 2                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             1                 1                 1                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`ncm-19`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            10                 9                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             4                 4                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             6                 6                 0                0
#>  7 (HN) Incr~             6                 6                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             1                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             1                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`ncf-20`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             4                 4                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             3                 3                 0                0
#>  7 (HN) Incr~             2                 2                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             1                 1                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbm-21`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            11                11                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             2                 2                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             4                 4                 0                0
#>  7 (HN) Incr~             5                 5                 0                0
#>  8 (BF) Incr~             1                 1                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             0                 0                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             3                 3                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbf-22`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            14                14                 0                0
#>  2 (BF) Incr~             1                 1                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~            10                10                 0                0
#>  7 (HN) Incr~             9                 9                 0                0
#>  8 (BF) Incr~             1                 1                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             3                 3                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             1                 1                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             2                 2                 0                0
#> 15 (BF) Incr~             1                 1                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbf-23`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             8                 8                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             2                 2                 0                0
#>  8 (BF) Incr~             1                 1                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             1                 1                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbm-24`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             8                 8                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             2                 2                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             3                 3                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             1                 1                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             1                 1                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`ncm-13`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             1                 1                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             0                 0                 0                0
#>  7 (HN) Incr~             0                 0                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             0                 0                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             1                 1                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbf-14`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            14                14                 0                0
#>  2 (BF) Incr~             3                 3                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~            13                13                 0                0
#>  7 (HN) Incr~             5                 5                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             1                 1                 0                0
#> 10 (IEA) Inc~             3                 3                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             2                 2                 0                0
#> 14 (RW) Incr~             2                 2                 0                0
#> 15 (BF) Incr~             1                 1                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbf-15`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             8                 8                 0                0
#>  2 (BF) Incr~             3                 3                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             4                 4                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             1                 1                 0                0
#> 11 (RW) Chil~             1                 1                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`ncf-16`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             9                 9                 0                0
#>  2 (BF) Incr~             4                 4                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             2                 2                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbf-17`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            10                10                 0                0
#>  2 (BF) Incr~             1                 1                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             2                 2                 0                0
#>  5 (IEA) Inc~             1                 1                 0                0
#>  6 (IEA) Inc~             8                 8                 0                0
#>  7 (HN) Incr~             4                 4                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             1                 1                 0                0
#> 10 (IEA) Inc~             5                 5                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbm-18`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             9                 9                 0                0
#>  2 (BF) Incr~             2                 2                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             9                 9                 0                0
#>  7 (HN) Incr~             3                 3                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             1                 1                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             1                 1                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjf-30`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             6                 6                 0                0
#>  2 (BF) Incr~             2                 2                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             6                 6                 0                0
#>  7 (HN) Incr~             2                 2                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             2                 2                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             1                 1                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`mjm-4`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             5                 4                 1                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             4                 3                 1                0
#>  7 (HN) Incr~             2                 2                 1                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             0                 0                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbf-2`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             8                 4                 0                4
#>  2 (BF) Incr~             1                 1                 0                1
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             0                 0                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             7                 4                 0                3
#>  7 (HN) Incr~             3                 3                 0                3
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             1                 0                 0                1
#> 10 (IEA) Inc~             2                 1                 0                2
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             0                 0                 0                0
#> 13 (IEA) Inc~             1                 0                 0                1
#> 14 (RW) Incr~             1                 0                 0                1
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $`nbm-3`
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~             4                 4                 0                0
#>  2 (BF) Incr~             0                 0                 0                0
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~             1                 1                 0                0
#>  5 (IEA) Inc~             0                 0                 0                0
#>  6 (IEA) Inc~             4                 4                 0                0
#>  7 (HN) Incr~             2                 2                 0                0
#>  8 (BF) Incr~             0                 0                 0                0
#>  9 (IEA) Inc~             0                 0                 0                0
#> 10 (IEA) Inc~             0                 0                 0                0
#> 11 (RW) Chil~             0                 0                 0                0
#> 12 (BF) Incr~             1                 1                 0                0
#> 13 (IEA) Inc~             0                 0                 0                0
#> 14 (RW) Incr~             0                 0                 0                0
#> 15 (BF) Incr~             0                 0                 0                0
#> 16 (RW) Incr~             0                 0                 0                0
#> 17 (IEA) Inc~             0                 0                 0                0
#> 18 (IEA) Inc~             0                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#> 
#> $summary
#> # A tibble: 18 x 7
#>    row_names  `All sources` `(IEA) Social Ca~ `(IEA) Social Ca~ `(IEA) Social C~
#>    <chr>              <dbl>             <dbl>             <dbl>            <dbl>
#>  1 All targe~            28                28                 4                1
#>  2 (BF) Incr~            12                12                 2                1
#>  3 (BF) Expe~             0                 0                 0                0
#>  4 (IEA) Inc~            16                16                 2                0
#>  5 (IEA) Inc~             1                 1                 0                0
#>  6 (IEA) Inc~            27                27                 4                1
#>  7 (HN) Incr~            27                27                 4                1
#>  8 (BF) Incr~             4                 4                 0                0
#>  9 (IEA) Inc~             4                 3                 0                1
#> 10 (IEA) Inc~            22                22                 1                1
#> 11 (RW) Chil~             1                 1                 0                0
#> 12 (BF) Incr~             7                 7                 0                0
#> 13 (IEA) Inc~            11                10                 0                1
#> 14 (RW) Incr~            11                10                 1                1
#> 15 (BF) Incr~             4                 4                 0                0
#> 16 (RW) Incr~             1                 0                 0                0
#> 17 (IEA) Inc~             1                 0                 0                0
#> 18 (IEA) Inc~             1                 0                 0                0
#> # ... with 2 more variables:
#> #   (IEA) Social Cash Transfer not working (Gov) <dbl>,
#> #   (IEA) Cash unreliable [N] <dbl>
#>