Find factors

Fields may be from the original data and/or fields created by the app, e.g. `n` (frequency). When field is 'label', 'value' can contain a vector of search terms separated by ' OR '.

Text searches are case-insensitive.

pipe_find_factors(graf, field = NULL, value, operator = NULL, up = 0, down = 0)

Arguments

field
Field (column, variable) to search

value
Value to search for

operator
c('contains','notcontains','=','notequals','greater','less','starts','ends'). How to search.

up
integer. Default is 0.

down
integer. Default is 0.

Value

A tidymap containing only matching factors; if `up`!=0 then also factors this number of steps upstream of the matching factors are also included and likewise for `down`!=0. The links are filtered correspondingly to return only the "ego network" i.e. links between the returned factors.

If operator and field are both NULL, the value is treated as a simple search string for the field `label`.

Examples

pipe_find_factors(cashTransferMap,NULL,"Cash")

#> # A tbl_graph: 7 nodes and 6 edges
#> #
#> # A directed acyclic multigraph with 3 components
#> #
#> # Node Data: 7 x 6 (active)
#>      id label                                    factor_note found upvec downvec
#>   <int> <chr>                                    <chr>       <lgl> <lgl> <lgl>  
#> 1     5 (IEA) Social Cash Transfer (OrgX) [E]    NA          TRUE  TRUE  TRUE   
#> 2     6 (IEA) Social Cash Transfer (Gov) [I]     NA          TRUE  TRUE  TRUE   
#> 3    29 (IEA) Social Cash Transfer (Oranisation~ NA          TRUE  TRUE  TRUE   
#> 4    31 (IEA) Social Cash Transfer not working ~ NA          TRUE  TRUE  TRUE   
#> 5   100 (IEA) Cash shared between villagers [N]  NA          TRUE  TRUE  TRUE   
#> 6   101 (IEA) Limited impact of Social Cash Tra~ NA          TRUE  TRUE  TRUE   
#> # ... with 1 more row
#> #
#> # Edge Data: 6 x 18
#>    from    to statement_id quote note  label attribution definition
#>   <int> <int>        <int> <chr> <chr> <chr> <chr>       <chr>     
#> 1     4     5            5 She ~ ""    ""    NA          NA        
#> 2     4     7          181 With~ ""    ""    NA          NA        
#> 3     4     7          198 Impo~ ""    ""    NA          NA        
#> # ... with 3 more rows, and 10 more variables: project_attribution <chr>,
#> #   valence_attribution <chr>, strength <dbl>, neg_from <int>, neg_to <int>,
#> #   hashtags <chr>, id <int>, actualisation <dbl>, certainty <dbl>,
#> #   weight <dbl>

pipe_find_factors(cashTransferMap,field="label",value="Cash",operator="contains")

#> # A tbl_graph: 7 nodes and 6 edges
#> #
#> # A directed acyclic multigraph with 3 components
#> #
#> # Node Data: 7 x 6 (active)
#>      id label                                    factor_note found upvec downvec
#>   <int> <chr>                                    <chr>       <lgl> <lgl> <lgl>  
#> 1     5 (IEA) Social Cash Transfer (OrgX) [E]    NA          TRUE  TRUE  TRUE   
#> 2     6 (IEA) Social Cash Transfer (Gov) [I]     NA          TRUE  TRUE  TRUE   
#> 3    29 (IEA) Social Cash Transfer (Oranisation~ NA          TRUE  TRUE  TRUE   
#> 4    31 (IEA) Social Cash Transfer not working ~ NA          TRUE  TRUE  TRUE   
#> 5   100 (IEA) Cash shared between villagers [N]  NA          TRUE  TRUE  TRUE   
#> 6   101 (IEA) Limited impact of Social Cash Tra~ NA          TRUE  TRUE  TRUE   
#> # ... with 1 more row
#> #
#> # Edge Data: 6 x 18
#>    from    to statement_id quote note  label attribution definition
#>   <int> <int>        <int> <chr> <chr> <chr> <chr>       <chr>     
#> 1     4     5            5 She ~ ""    ""    NA          NA        
#> 2     4     7          181 With~ ""    ""    NA          NA        
#> 3     4     7          198 Impo~ ""    ""    NA          NA        
#> # ... with 3 more rows, and 10 more variables: project_attribution <chr>,
#> #   valence_attribution <chr>, strength <dbl>, neg_from <int>, neg_to <int>,
#> #   hashtags <chr>, id <int>, actualisation <dbl>, certainty <dbl>,
#> #   weight <dbl>

pipe_find_factors(cashTransferMap,field="id",value=10,operator="greater")

#> # A tbl_graph: 117 nodes and 806 edges
#> #
#> # A directed multigraph with 11 components
#> #
#> # Node Data: 117 x 6 (active)
#>      id label                                   factor_note found upvec downvec
#>   <int> <chr>                                   <chr>       <lgl> <lgl> <lgl>  
#> 1    11 (RW) Part of Organisation X project [E] NA          TRUE  TRUE  TRUE   
#> 2    12 (IEA) Government HSA                    NA          TRUE  TRUE  TRUE   
#> 3    13 (HN) More food                          NA          TRUE  TRUE  TRUE   
#> 4    14 (RW) Husband returned to household      NA          TRUE  TRUE  TRUE   
#> 5    15 (BF) Purchased livestock                NA          TRUE  TRUE  TRUE   
#> 6    16 (BF) Demonstration plot [E]             NA          TRUE  TRUE  TRUE   
#> # ... with 111 more rows
#> #
#> # Edge Data: 806 x 18
#>    from    to statement_id quote note  label attribution definition
#>   <int> <int>        <int> <chr> <chr> <chr> <chr>       <chr>     
#> 1    39    40            2 "The~ ""    ""    NA          NA        
#> 2    40    41            2 "Suc~ ""    ""    NA          NA        
#> 3    49    50            3 "The~ ""    ""    NA          NA        
#> # ... with 803 more rows, and 10 more variables: project_attribution <chr>,
#> #   valence_attribution <chr>, strength <dbl>, neg_from <int>, neg_to <int>,
#> #   hashtags <chr>, id <int>, actualisation <dbl>, certainty <dbl>,
#> #   weight <dbl>

pipe_find_factors(cashTransferMap,NULL,"purchase OR buy")

#> # A tbl_graph: 3 nodes and 0 edges
#> #
#> # A rooted forest with 3 trees
#> #
#> # Node Data: 3 x 6 (active)
#>      id label                                    factor_note found upvec downvec
#>   <int> <chr>                                    <chr>       <lgl> <lgl> <lgl>  
#> 1    15 (BF) Purchased livestock                 NA          TRUE  TRUE  TRUE   
#> 2    37 (BF) Organisation A failed to buy soya   NA          TRUE  TRUE  TRUE   
#> 3    66 (IEA) No longer sell assets to buy food~ NA          TRUE  TRUE  TRUE   
#> #
#> # Edge Data: 0 x 18
#> # ... with 18 variables: from <int>, to <int>, statement_id <int>, quote <chr>,
#> #   note <chr>, label <chr>, attribution <chr>, definition <chr>,
#> #   project_attribution <chr>, valence_attribution <chr>, strength <dbl>,
#> #   neg_from <int>, neg_to <int>, hashtags <chr>, id <int>,
#> #   actualisation <dbl>, certainty <dbl>, weight <dbl>