Merge statements into links

Merge statements into links

pipe_merge_statements(graf)

Arguments

graf
A tidymap representing a causal map. A tidymap is a tidygraph, which consists of a table of edges linked to a table of nodes, with an optional additional table of statements. In this package, nodes are called `factors` and edges are called `links.`

Value

A tidymap in which columns from the statements table are merged into the links table.

Examples

cashTransferMap %>% pipe_merge_statements() %>% pipe_find_links(field="text",value="women",operator="contains")

#> # A tbl_graph: 127 nodes and 46 edges
#> #
#> # A directed acyclic multigraph with 98 components
#> #
#> # Node Data: 127 x 3 (active)
#>      id label                                     factor_note
#>   <int> <chr>                                     <chr>      
#> 1     1 (RW) Birth                                NA         
#> 2     2 (HN) WASH Information [E]                 NA         
#> 3     3 (HN) Malaria prevention advice [E]        NA         
#> 4     4 (BF) Agricultural training and advice [E] NA         
#> 5     5 (IEA) Social Cash Transfer (OrgX) [E]     NA         
#> 6     6 (IEA) Social Cash Transfer (Gov) [I]      NA         
#> # ... with 121 more rows
#> #
#> # Edge Data: 46 x 90
#>    from    to statement_id quote note  label attribution definition
#>   <int> <int>        <int> <chr> <chr> <chr> <chr>       <chr>     
#> 1     5    76           13 The ~ ""    ""    NA          NA        
#> 2    10    78           13 The ~ ""    ""    NA          NA        
#> 3    10    78          330 The ~  NA   ""    NA          NA        
#> # ... with 43 more rows, and 82 more variables: project_attribution <chr>,
#> #   valence_attribution <chr>, strength <dbl>, neg_from <int>, neg_to <int>,
#> #   hashtags <chr>, id.x <int>, actualisation <dbl>, certainty <dbl>,
#> #   weight <dbl>, text <chr>, statement_note <chr>, `#QuestionID` <chr>,
#> #   `#SourceID` <chr>, `*amount saved -- F6_Overall, how has the amount you
#> #   save as a household changed over the last year (net of any debts)?` <chr>,
#> #   `*assets -- F7_Overall, do you feel the combined total value of all your
#> #   assets has gone up or down over the period?` <chr>, `*farming --
#> #   C2_Overall, how have your farming activities changed in the last
#> #   year?` <chr>, `*food amount -- E2_Overall, has the amount of food you eat
#> #   increased or decreased in the last year?` <chr>, `*food quality --
#> #   E4_Overall, has the quality of your familys food got better or worse in the
#> #   last year?` <chr>, `*health family -- B2_Overall, how has the health of
#> #   your family changed in this time?` <chr>, `*money - control -- F4_Overall,
#> #   has the control you have personally over how you spend money changed over
#> #   the last year?` <chr>, `*money earned -- D2_Overall, how has the amount of
#> #   money your household earns changed in the last year?` <chr>, `*purchasing
#> #   power -- F2_Overall, how has the amount you can buy as a household changed
#> #   over the last year?` <chr>, `*wellbeing - household -- H3_Overall, taking
#> #   all things into account, how do you think the wellbeing of your household
#> #   has changed during this period?` <chr>, `*wellbeing - respondent --
#> #   H1_Overall, taking all things into account, how do you think your own
#> #   wellbeing has changed during this period?` <chr>, `1. age -- A1a_What is
#> #   the age of the main respondent?` <int>, `1. education -- A1b_What is the
#> #   education of the main respondent?` <chr>, `1. respondent_1` <lgl>, `1.
#> #   Sex` <chr>, `1. Type` <chr>, `1.Performance` <chr>, `2. Village` <chr>,
#> #   `A1_Focus Group Code` <chr>, `A1_Household code` <chr>,
#> #   A10_Recording <lgl>, `A10_To make sure our record of the interview is
#> #   accurate we would like to make an audio-recording of the interview. Is the
#> #   group happy for us to make this recording?` <chr>, `A11_Most of our
#> #   questions refer to what has happened in the last year. Can you think back
#> #   to something important that happened to you 12 months ago? What was it?
#> #   Please answer questions below by thinking back to that time.` <chr>,
#> #   `A11_To make sure our record of the interview is accurate we would like to
#> #   make an audio-recording of the interview. Are you (both) happy for us to
#> #   make this recording?` <chr>, `A12_Most of our questions refer to what has
#> #   happened in the last year. Can you think back to something important that
#> #   happened to you 12 months ago? What was it? Please answer questions below
#> #   by thinking back to that time.` <chr>, `A14_Please note down the gender of
#> #   the head of the household` <chr>, `A15_Does anyone in the household have a
#> #   chronic illness or disability?` <chr>, `A1c_What is the gender of the main
#> #   respondent?` <chr>, `A1d_What is the household size (residents >
#> #   6months)?` <int>, `A1e_How many under 16s in the household?` <int>,
#> #   `A1f_How many under 16s in full time education in the household?` <int>,
#> #   `A2_Name of village` <chr>, `A2_Type of focus group` <chr>, `A3_Name of
#> #   facilitator` <chr>, `A3_Name of interviewer` <chr>, `A4_Date of
#> #   interview` <chr>, `A4_Location of meeting` <chr>, `A5_Date of
#> #   meeting` <chr>, `A6_End time of interview (hh:mm)` <dbl>, `A6_Start
#> #   time` <dbl>, `A6a_Duration of interview (in minutes)` <dbl>, `A7_Are you
#> #   willing to be interviewed?` <chr>, `A7_End time` <dbl>, `A7a_Duration of
#> #   interview (in minutes)` <dbl>, `A8_Number of participants` <int>, `A9_IF
#> #   YES: we would like to hear your personal answers to our questions. Are you
#> #   willing to be interviewed alone?` <chr>, `A9_Number of participants already
#> #   interviewed in hh interviews` <int>, District <chr>, `Domain Attributed
#> #   IH` <chr>, `H1_Overall, taking all things into account, how do you think
#> #   your wellbeing - and that of people like you - has changed during this
#> #   period? Why is this?` <chr>, `I1i_Important links to organisations,
#> #   rank` <chr>, `I1j_Important links to organisations, rank` <chr>,
#> #   `I1k_Important links to organisations, rank` <chr>, `I1l_Important links to
#> #   organisations, rank` <chr>, `I1m_Important links to organisations,
#> #   rank` <chr>, `I1n_Important links to organisations, rank` <chr>,
#> #   `I1o_Important links to organisations, rank` <chr>, `I1p_Important links to
#> #   organisations, rank` <chr>, `I2_Dont forget to thank the interviewee for
#> #   their participation, and invite them to ask you any questions they have.
#> #   Note down what these are, and if they prompt issues relevant to the
#> #   research that you would like to share.NOTES:` <chr>, `I2_Dont forget to
#> #   thank the participants, and invite them to ask you any questions they have.
#> #   Note down what these are, and if they prompt issues relevant to the
#> #   research that you would like to share.NOTES:` <chr>, `I3_Please also note
#> #   down any observations you have about the respondent and the interview
#> #   process (e.g. were they relaxed, were there distractions, or do you think
#> #   their answers influenced by other people present at the interview?)` <chr>,
#> #   id.y <chr>, QID <chr>, X6 <lgl>, X7 <lgl>, X8 <lgl>, xQID <chr>,
#> #   found <lgl>

cashTransferMap %>% pipe_find_statements(field="text",value="women",operator="contains")

#> # A tbl_graph: 127 nodes and 46 edges
#> #
#> # A directed acyclic multigraph with 98 components
#> #
#> # Node Data: 127 x 3 (active)
#>      id label                                     factor_note
#>   <int> <chr>                                     <chr>      
#> 1     1 (RW) Birth                                NA         
#> 2     2 (HN) WASH Information [E]                 NA         
#> 3     3 (HN) Malaria prevention advice [E]        NA         
#> 4     4 (BF) Agricultural training and advice [E] NA         
#> 5     5 (IEA) Social Cash Transfer (OrgX) [E]     NA         
#> 6     6 (IEA) Social Cash Transfer (Gov) [I]      NA         
#> # ... with 121 more rows
#> #
#> # Edge Data: 46 x 18
#>    from    to statement_id quote note  label attribution definition
#>   <int> <int>        <int> <chr> <chr> <chr> <chr>       <chr>     
#> 1     5    76           13 The ~ ""    ""    NA          NA        
#> 2    10    78           13 The ~ ""    ""    NA          NA        
#> 3    10    78          330 The ~  NA   ""    NA          NA        
#> # ... with 43 more rows, and 10 more variables: project_attribution <chr>,
#> #   valence_attribution <chr>, strength <dbl>, neg_from <int>, neg_to <int>,
#> #   hashtags <chr>, id <int>, actualisation <dbl>, certainty <dbl>,
#> #   weight <dbl>