Find factors

Fields may be from the original data and/or fields created by the app, e.g. `n` (frequency). When field is 'label', 'value' can contain a vector of search terms separated by ' OR '.

Text searches are case-insensitive.

Usage

pipe_find_factors(
  graf,
  field = "label",
  value,
  operator = "contains",
  up = 1,
  down = 1,
  remove_isolated = F,
  highlight_only = F
)

Arguments

field

Field (column, variable) to search

value

Value to search for

operator

c('contains','notcontains','=','notequals','greater','less','starts','ends'). How to search.

up

integer. Default is 0.

down

integer. Default is 0.

Value

A mapfile containing only matching factors; if `up`!=0 then also factors this number of steps upstream of the matching factors are also included and likewise for `down`!=0. The links are filtered correspondingly to return only the "ego network" i.e. links between the returned factors. If operator and field are both NULL, the value is treated as a simple search string for the field `label`.

Examples

pipe_find_factors(example2,field="label",value="Flood",operator="contains")
#> $factors
#> # A tibble: 8 x 17
#>   label         id    factor_note factor_id0 factor_id factor_memo factor_map_id
#>   <chr>         <chr> <lgl>            <int>     <int> <chr>               <int>
#> 1 "Better floo~ 6     NA                   1         1 ""                      1
#> 2 "Damage to B~ 4     NA                   3         3 ""                      1
#> 3 "Damage to P~ 3     NA                   4         4 ""                      1
#> 4 "Flood risk ~ 7     NA                   5         5 ""                      1
#> 5 "Flooding \U~ 2     NA                   6         6 ""                      1
#> 6 "Funds from ~ 5     NA                   7         7 ""                      1
#> 7 "High rainfa~ 1     NA                   8         8 ""                      1
#> 8 "Increased i~ 8     NA                   9         9 ""                      1
#> # ... with 10 more variables: betweenness <dbl>, betweenness_rank <dbl>,
#> #   in_degree <dbl>, out_degree <dbl>, frequency <dbl>, size <dbl>,
#> #   found <lgl>, found_type <chr>, map_id <int>, is_flipped <lgl>
#> 
#> $links
#> # A tibble: 6 x 59
#>   from_label  to_label  quote  bundle   from    to note  link_id0 link_id weight
#>   <chr>       <chr>     <chr>  <chr>   <int> <int> <lgl>    <int>   <int>  <int>
#> 1 "High rain~ "Floodin~ Recen~ "High ~     8     6 NA           1       1      1
#> 2 "Flooding ~ "Damage ~ The f~ "Flood~     6     4 NA           2       2      1
#> 3 "Flooding ~ "Damage ~ The f~ "Flood~     6     3 NA           3       3      1
#> 4 "Funds fro~ "Better ~ Funds~ "Funds~     7     1 NA           4       4      1
#> 5 "Better fl~ "Flood r~ bette~ "Bette~     1     5 NA           5       5      1
#> 6 "Flood ris~ "Increas~ reduc~ "Flood~     5     9 NA           6       6      1
#> # ... with 49 more variables: strength <int>, certainty <int>,
#> #   from_flipped <lgl>, to_flipped <lgl>, link_label <chr>, hashtag <chr>,
#> #   link_memo <chr>, link_map_id <int>, s.text <chr>, s.statement_note <lgl>,
#> #   s.#QuestionID <lgl>, s.#SourceID <lgl>, s.statement_memo <chr>,
#> #   s.source_id <chr>, s.question_id <chr>, s.statement_map_id <dbl>,
#> #   s.factor_id <int>, s.r.source_memo <chr>, s.r.source_map_id <dbl>,
#> #   s.q.question_text <chr>, s.q.question_memo <chr>, ...
#> 
#> $statements
#> # A tibble: 2 x 21
#>   statement_id text      statement_note `#QuestionID` `#SourceID` statement_memo
#>          <dbl> <chr>     <lgl>          <lgl>         <lgl>       <chr>         
#> 1            1 Welcome ~ NA             NA            NA          st memo       
#> 2            2 Rising s~ NA             NA            NA          NA            
#> # ... with 15 more variables: statement_map_id <dbl>, factor_id <int>,
#> #   r.source_memo <chr>, r.source_map_id <dbl>, q.question_text <chr>,
#> #   q.question_memo <chr>, q.question_map_id <dbl>, label <dbl>,
#> #   source_id <chr>, source_memo <chr>, source_map_id <dbl>, question_id <chr>,
#> #   question_text <chr>, question_memo <chr>, question_map_id <dbl>
#> 
#> $sources
#> # A tibble: 1 x 3
#>   source_id source_memo source_map_id
#>   <chr>     <chr>               <dbl>
#> 1 1         ooh                     1
#> 
#> $questions
#> # A tibble: 1 x 4
#>   question_id question_text   question_memo question_map_id
#>   <chr>       <chr>           <chr>                   <dbl>
#> 1 1           global question q memo                      1
#> 
#> $settings
#> # A tibble: 1 x 3
#>   setting_id        value map_id
#>   <chr>             <chr> <chr> 
#> 1 background_colour NA    1     
#> 
#> attr(,"class")
#> [1] "mapfile" "list"   
#> attr(,"meta")
#> attr(,"meta")$pipe_find_factors
#> attr(,"meta")$pipe_find_factors[[1]]
#> pipe_find_factors
#> 
#> attr(,"meta")$pipe_find_factors$graf
#> example2
#> 
#> attr(,"meta")$pipe_find_factors$field
#> [1] "label"
#> 
#> attr(,"meta")$pipe_find_factors$value
#> [1] "Flood"
#> 
#> attr(,"meta")$pipe_find_factors$operator
#> [1] "contains"
#> 
#> 
pipe_find_factors(example2,value="Flood") %>% make_interactive_map

pipe_find_factors(example2,value="Flood")
#> $factors
#> # A tibble: 8 x 17
#>   label         id    factor_note factor_id0 factor_id factor_memo factor_map_id
#>   <chr>         <chr> <lgl>            <int>     <int> <chr>               <int>
#> 1 "Better floo~ 6     NA                   1         1 ""                      1
#> 2 "Damage to B~ 4     NA                   3         3 ""                      1
#> 3 "Damage to P~ 3     NA                   4         4 ""                      1
#> 4 "Flood risk ~ 7     NA                   5         5 ""                      1
#> 5 "Flooding \U~ 2     NA                   6         6 ""                      1
#> 6 "Funds from ~ 5     NA                   7         7 ""                      1
#> 7 "High rainfa~ 1     NA                   8         8 ""                      1
#> 8 "Increased i~ 8     NA                   9         9 ""                      1
#> # ... with 10 more variables: betweenness <dbl>, betweenness_rank <dbl>,
#> #   in_degree <dbl>, out_degree <dbl>, frequency <dbl>, size <dbl>,
#> #   found <lgl>, found_type <chr>, map_id <int>, is_flipped <lgl>
#> 
#> $links
#> # A tibble: 6 x 59
#>   from_label  to_label  quote  bundle   from    to note  link_id0 link_id weight
#>   <chr>       <chr>     <chr>  <chr>   <int> <int> <lgl>    <int>   <int>  <int>
#> 1 "High rain~ "Floodin~ Recen~ "High ~     8     6 NA           1       1      1
#> 2 "Flooding ~ "Damage ~ The f~ "Flood~     6     4 NA           2       2      1
#> 3 "Flooding ~ "Damage ~ The f~ "Flood~     6     3 NA           3       3      1
#> 4 "Funds fro~ "Better ~ Funds~ "Funds~     7     1 NA           4       4      1
#> 5 "Better fl~ "Flood r~ bette~ "Bette~     1     5 NA           5       5      1
#> 6 "Flood ris~ "Increas~ reduc~ "Flood~     5     9 NA           6       6      1
#> # ... with 49 more variables: strength <int>, certainty <int>,
#> #   from_flipped <lgl>, to_flipped <lgl>, link_label <chr>, hashtag <chr>,
#> #   link_memo <chr>, link_map_id <int>, s.text <chr>, s.statement_note <lgl>,
#> #   s.#QuestionID <lgl>, s.#SourceID <lgl>, s.statement_memo <chr>,
#> #   s.source_id <chr>, s.question_id <chr>, s.statement_map_id <dbl>,
#> #   s.factor_id <int>, s.r.source_memo <chr>, s.r.source_map_id <dbl>,
#> #   s.q.question_text <chr>, s.q.question_memo <chr>, ...
#> 
#> $statements
#> # A tibble: 2 x 21
#>   statement_id text      statement_note `#QuestionID` `#SourceID` statement_memo
#>          <dbl> <chr>     <lgl>          <lgl>         <lgl>       <chr>         
#> 1            1 Welcome ~ NA             NA            NA          st memo       
#> 2            2 Rising s~ NA             NA            NA          NA            
#> # ... with 15 more variables: statement_map_id <dbl>, factor_id <int>,
#> #   r.source_memo <chr>, r.source_map_id <dbl>, q.question_text <chr>,
#> #   q.question_memo <chr>, q.question_map_id <dbl>, label <dbl>,
#> #   source_id <chr>, source_memo <chr>, source_map_id <dbl>, question_id <chr>,
#> #   question_text <chr>, question_memo <chr>, question_map_id <dbl>
#> 
#> $sources
#> # A tibble: 1 x 3
#>   source_id source_memo source_map_id
#>   <chr>     <chr>               <dbl>
#> 1 1         ooh                     1
#> 
#> $questions
#> # A tibble: 1 x 4
#>   question_id question_text   question_memo question_map_id
#>   <chr>       <chr>           <chr>                   <dbl>
#> 1 1           global question q memo                      1
#> 
#> $settings
#> # A tibble: 1 x 3
#>   setting_id        value map_id
#>   <chr>             <chr> <chr> 
#> 1 background_colour NA    1     
#> 
#> attr(,"class")
#> [1] "mapfile" "list"   
#> attr(,"meta")
#> attr(,"meta")$pipe_find_factors
#> attr(,"meta")$pipe_find_factors[[1]]
#> pipe_find_factors
#> 
#> attr(,"meta")$pipe_find_factors$graf
#> example2
#> 
#> attr(,"meta")$pipe_find_factors$value
#> [1] "Flood"
#> 
#> 
pipe_find_factors(example2,field="id",value=5,operator="greater")
#> Error in if (df$found %>% sum %>% ==0) {    graf <- (pipe_update_mapfile(graf, factors = graf$factors %>%         filter(F), links = graf$links %>% filter(F)))    return(graf %>% add_meta(meta))}: missing value where TRUE/FALSE needed